// reelease-ai-next/src/types/socialMedia.ts
export interface CountdownMore {
  days: number
  hrs: number
  mins: number
  isMoreThan24Hours: true
  secs?: never
}

export interface CountdownLess {
  days?: never
  hrs: number
  mins: number
  secs: number
  isMoreThan24Hours: false
}

export type CountdownResult = CountdownMore | CountdownLess | null